+++ packages/unlock-js/src/PublicLock/v10/purchaseKey.js
@@ -1,7 +1,7 @@
 import purchaseKeys from './purchaseKeys'
 
 /**
- * Purchase key function. calls the purchaseKeys function with the vlalue in array
+ * Purchase key function. calls the purchaseKeys function with the value in array
  */
 export default async function (
   {
@@ -12,6 +12,7 @@ export default async function (
     erc20Address,
     decimals,
     referrer,
+    recurringPayments, // nb of reccuring payments to approve
     data,
   },
   callback
@@ -23,6 +24,7 @@ export default async function (
       keyPrices: keyPrice ? [keyPrice] : null,
       referrers: referrer ? [referrer] : null,
       data: data ? [data] : null,
+      recurringPayments: recurringPayments ? [recurringPayments] : null,
       lockAddress,
       erc20Address,
       decimals,