+++ smart-contracts/hardhat.config.js
@@ -106,6 +106,7 @@ require('./tasks/release')
 require('./tasks/gov')
 require('./tasks/utils')
 require('./tasks/lock')
+require('./tasks/verify')
 
 /**
  * @type import('hardhat/config').HardhatUserConfig
@@ -115,7 +116,7 @@ module.exports = {
   etherscan,
   gasReporter: {
     currency: 'USD',
-    excludeContracts: ['Migrations', 'TestNoop'],
+    excludeContracts: ['TestNoop'],
     gasPrice: 5,
   },
   solidity: {