+++ unlock-app/src/hooks/useAddToNetwork.ts
@@ -1,5 +1,5 @@
-import { toast } from 'react-hot-toast'
 import { useContext, useEffect, useState } from 'react'
+import { ToastHelper } from '../components/helpers/toast.helper'
 import { ProviderContext } from '../contexts/ProviderContext'
 import { ConfigContext } from '../utils/withConfig'
 
@@ -36,7 +36,7 @@ export const useAddToNetwork = (account?: string) => {
     }
 
     const promise = provider.send('wallet_addEthereumChain', [params], account)
-    await toast.promise(promise, {
+    await ToastHelper.promise(promise, {
       loading: `Changing network to ${chainName}. Please Approve on your wallet.`,
       error: `Error in changing network to ${chainName}`,
       success: `Successfully changed network to ${chainName}`,