+++ packages/unlock-js/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes
 
+# 0.26.0
+
+- add support for `approveBeneficiary`
+- pass address as param to `getAllowance` instead of signer
+
 # 0.25.0
 
 - Better support for v10